+++ docs/source/change-log.rst
@@ -32,3 +32,9 @@ year-month-day
 2022-04-14
 
 * Added :py:meth:`~ipumspy.ddi.Codebook.get_all_types()` method to access all types of ddi codebook variables in an easy way.
+
+2022-04-20
+
+* Added parameter `string_pyarrow` to :py:meth:`~ipumspy.ddi.Codebook.get_all_types()` method. If this parameter is set to True is used in conjunction
+  with parameter `type_format`="pandas_type", then the string column dtype (pd.StringDtype()) is overriden with pd.StringDtype(storage="pyarrow"). Useful for
+  users who want to convert an IPUMS extract in csv format to parquet format.